Add files via upload
The biggest change was changing abilMod from a list to a function. Before, values in the list were set when the stats were randomized, but stayed constant if the stats were updated by race modifiers or other modifiers. As a function, abilMod now calculates the ability modifiers from the stats whenever it is called.